Extract XPath test utility for reuse
