+++ src/test/kotlin/io/hemin/wien/writer/namespace/NamespaceWriterTest.kt
@@ -7,6 +7,7 @@ import io.hemin.wien.documentFromResource
 import io.hemin.wien.dom.DomBuilderFactory
 import io.hemin.wien.dom.appendElement
 import io.hemin.wien.dom.findElementByName
+import io.hemin.wien.dom.findNodeByXPath
 import io.hemin.wien.model.Episode
 import io.hemin.wien.model.Podcast
 import io.hemin.wien.model.episode.anEpisode
@@ -18,14 +19,8 @@ import org.w3c.dom.Node
 import org.xmlunit.builder.DiffBuilder
 import org.xmlunit.builder.Input
 import org.xmlunit.diff.Diff
-import javax.xml.namespace.NamespaceContext
-import javax.xml.xpath.XPathConstants
-import javax.xml.xpath.XPathFactory
 
 internal abstract class NamespaceWriterTest {
-
-    private val xpath = XPathFactory.newDefaultInstance().newXPath()
-        .apply { namespaceContext = FeedNamespaceContext }
 
     private val documentBuilder = DomBuilderFactory.newDocumentBuilder()
 
@@ -39,7 +34,7 @@ internal abstract class NamespaceWriterTest {
 
     private fun expectedNode(nodePath: String, resourcePath: String): Node {
         val document = documentFromResource(resourcePath)
-        val node = xpath.evaluate(nodePath, document, XPathConstants.NODE) as? Node
+        val node = document.findNodeByXPath(nodePath)
         requireNotNull(node) { "The node with XPath '$nodePath' could not be found in resource '$resourcePath'" }
         return node
     }
@@ -109,23 +104,5 @@ internal abstract class NamespaceWriterTest {
     protected fun Element.createItemElement(): Element {
         require(namespaceURI == null && nodeName == "channel") { " Only a <channel> can contain items" }
         return appendElement("item")
-    }
-}
-
-private object FeedNamespaceContext : NamespaceContext {
-
-    override fun getNamespaceURI(prefix: String?): String? {
-        if (prefix == null) return null
-        return FeedNamespace.values().find { it.prefix == prefix }?.uri
-    }
-
-    override fun getPrefix(namespaceURI: String?): String? {
-        if (namespaceURI == null) return null
-        return FeedNamespace.values().find { it.uri == namespaceURI }?.prefix
-    }
-
-    override fun getPrefixes(namespaceURI: String?): Iterator<String> {
-        val prefix = getPrefix(namespaceURI) ?: return emptyList<String>().iterator()
-        return listOf(prefix).iterator()
     }
 }
